fix: remove unnecessary executeOperationsWithUniversalResolver helper
